% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_bars.R
\name{categorical_bars}
\alias{categorical_bars}
\title{List of Bar Plots}
\usage{
categorical_bars(
  data_frame,
  cat_vars,
  target_col,
  title_size = 30,
  axis_size = 35
)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{cat_vars}{A vector containing the string name(s) of each \strong{categorical} variable to plot on the x-axis.}

\item{target_col}{Column of interest to plot each variable against (object).}

\item{title_size}{Size of each plot's title (Default = 30).}

\item{axis_size}{Size of each plot's axes (Default = 35).}
}
\value{
A list containing \code{ggplot2} object(s) (bar plots).
}
\description{
Create a list of bar plots for each categorical variable in a data frame.
The proportions for each categorical variable are grouped by a different fill
color.
}
\examples{
\dontrun{
  # Generate bar plots for categorical variables from the mtcars dataset
  bar_plots <- categorical_bars(
    data_frame = mtcars,
    cat_vars = c("cyl", "gear"),
    target_col = "am",
    title_size = 25,
    axis_size = 20
  )
  # Display the first bar plot
  print(bar_plots[["cyl"]])
}

}
