% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/category_target.R
\name{category_target}
\alias{category_target}
\title{Counts and Proportions of Categorical Variables}
\usage{
category_target(data_frame, cat_var)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{cat_var}{A categorical variable (object) within data_frame.}
}
\value{
A data frame with 1 rows per category and 2 columns:
\itemize{
\item Count: Number of occurrances of each category in cat_var.
\item Proportion: Proportion of each category relative to the total number of
rows in the data frame.
}
}
\description{
Given an input dataframe with a categorical variable, return the number and
proportion of instances with each category value, in alphabetical
(for characters and logical) or numerical (for integers) order
}
\examples{
\dontrun{
  category_target(ToothGrowth, ToothGrowth$supp)
}
}
