% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cramer_chi_results.R
\name{cramer_chi_results}
\alias{cramer_chi_results}
\title{Table of Chi-Squared Test and Cramer's V Results}
\usage{
cramer_chi_results(df, categorical_vars, target_col)
}
\arguments{
\item{df}{A data frame or data frame extension (e.g. a tibble).}

\item{categorical_vars}{A vector containing the string name(s) of each \strong{categorical} variable in the data frame.}

\item{target_col}{A string specifying the categorical variable against which chi-squared tests will be performed.}
}
\value{
Data frame with 1 row per variable and 7 columns:
\itemize{
\item Variable: Name of categorical variable.
\item Statistic: Chi-squared test statistic.
\item DF: Degrees of freedom.
\item p_value: p-value from chi-squared test.
\item Expected_Min: Minimum expected value.
\item Expected_Max: Maximum expected value.
\item CramersV: Cramer's V statistic.
}
}
\description{
Runs chi-squared tests and calculates Cramer's V independently for each variable
in a given data frame
}
\examples{
\dontrun{
  cramer_chi_results(mtcars, c("cyl", "gear"), "mpg")
}

}
