% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count_type.R
\name{na_count_type}
\alias{na_count_type}
\title{Summarise Data Frame Columns}
\usage{
na_count_type(data_frame)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}
}
\value{
Data frame with 1 row per variable and 3 columns:
\itemize{
\item NA_Count: Number of "NA" values within each variable.
\item Distinct_Count: Number of distinct values for each variable.
\item Current_Data_Type: Current data type for each variable.
}
}
\description{
Given an input data frame, this function checks for the number of missing values
(NA), the number of distinct values, and the data type of each variable. \code{NULL}
values will be treated as a column of \code{NA}
}
\examples{
\dontrun{
  # Summarize columns in the mtcars dataset
  na_count_type(mtcars)
}

}
