% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_grid.R
\name{plots_grid}
\alias{plots_grid}
\title{Grid of Plots}
\usage{
plots_grid(bar_plots, num_cols = 3)
}
\arguments{
\item{bar_plots}{A list of \code{ggplot2} bar plot objects.}

\item{num_cols}{The number of columns in the grid (Default = 3).}
}
\value{
A \code{patchwork} / \code{ggplot2} object showing all input plots arranged.
}
\description{
Facet a list of \code{ggplot2} bar plot objects into a single
grid layout using \code{patchwork}. The number of columns in the grid can be
specified with \code{num_cols}.
}
\examples{
\dontrun{
  # Generate example bar plots
  bar_plots <- categorical_bars(
    data_frame = mtcars,
    cat_vars = c("cyl", "gear"),
    target_col = "am"
    )
    
  # Combine plots into a grid
  grid_plot <- plots_grid(bar_plots, num_cols = 2)
}

}
