% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantitative_density.R
\name{quantitative_density}
\alias{quantitative_density}
\title{List of Density Plots}
\usage{
quantitative_density(
  data_frame,
  noncat_vars,
  target_col,
  title_size = 35,
  axis_size = 30
)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{noncat_vars}{A vector containing the string name(s) of each \strong{non-categorical} variable to plot on the x-axis.}

\item{target_col}{Column of interest to plot each variable against (object).}

\item{title_size}{Size of each plot's title (Default = 30).}

\item{axis_size}{Size of each plot's axes (Default = 35).}
}
\value{
A list containing \code{ggplot2} object(s) (density plots).
}
\description{
Create a list of quantitative density plots for each non-categorical variable
in a data frame. The density area for each variable are mapped to a different
fill color.
}
\examples{
\dontrun{
  # Generate density plots for non-categorical variables from the mtcars dataset
  density_plots <- quantitative_density(
    data_frame = mtcars,
    noncat_vars = c("mpg", "hp", "wt"),
    target_col = "am",
    title_size = 25,
    axis_size = 20
  )
  # Display the first density plot
  print(density_plots[["mpg"]])
}
}
