% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_plot.R
\name{roc_plot}
\alias{roc_plot}
\title{Creates and saves an ROC curve plot}
\usage{
roc_plot(
  model_outputs,
  true_class,
  predicted_probs,
  roc_auc_value,
  output_path
)
}
\arguments{
\item{model_outputs}{A data frame or tibble containing the model outputs.}

\item{true_class}{A character string specifying the column with actual class labels.}

\item{predicted_probs}{A character string specifying the column with predicted probabilities.}

\item{roc_auc_value}{A numeric value representing the AUC from model evaluation.}

\item{output_path}{A character string specifying the file path to save the ROC plot.}
}
\value{
A \code{ggplot2} object of the ROC curve.
}
\description{
This function creates an ROC curve plot for a given classification model and
saves it to the specified file path.
}
\examples{
\dontrun{
# Assuming you have model outputs, true labels, predicted probabilities, and AUC value
roc_plot(lasso_model_outputs, "Diabetes_binary", ".pred_1",
         lasso_metrics$.estimate[lasso_metrics$.metric == "roc_auc"],
         "roc_curve_plot.png")
}
}
